+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 - BLD: warning free builds in any configuration
+- BUG: fix time_to_next_pixel computation with branchless (and without fma)
+       This bug was introduced in rlic 0.5.2
 
 ## 0.5.2 - 2025-11-14